vaultseed/requirements.txt
hvac
pyyaml
names
pymongo
jinja2
